delete identifier value object and add static constructors

+++ api/src/Denarius.Application/UseCases/Accounts/DeleteAccountUseCase.cs
@@ -3,7 +3,6 @@ using Denarius.Application.Exceptions;
 using Denarius.Application.Interfaces.Accounts;
 using Denarius.Application.Results;
 using Denarius.Domain.Interfaces;
-using Denarius.Domain.ValueObjects;
 
 namespace Denarius.Application.UseCases.Accounts;
 
@@ -11,18 +10,16 @@ internal class DeleteAccountUseCase(IAccountRepository accountRepository, ITrans
 {
     public async Task<AccountResult> Execute(IdCommand command)
     {
-        var id = new Identifier(command.Id);
-        var account = await accountRepository.FindByIdAsync(id);
+        var account = await accountRepository.FindByIdAsync(command.Id);
 
-        if (account is null)
-            throw new NotFoundException("Account not found.");
+        if (account is null) throw new NotFoundException("Account not found.");
 
         var transactions = (await transactionRepository.FindByAccountAsync(account.Id)).ToList();
-        
-        if (transactions.Count > 0)
-            transactionRepository.DeleteBatch(transactions);
+
+        if (transactions.Count > 0) transactionRepository.DeleteBatch(transactions);
 
         accountRepository.Delete(account);
+        await accountRepository.SaveAsync();
 
         return new AccountResult(account);
     }
